Animation

+++ src/Waiter.cpp
@@ -4,7 +4,7 @@ Waiter::Waiter(int interval)
 : interval(interval) {
 }
 
-bool Waiter::isReady(int currentTime) {
+bool Waiter::isReady(unsigned long currentTime) {
     bool result = false;
 
     if (currentTime >= previousTime + interval) {
